requirements.txt
django >= 2.2
python-fsutil >= 0.7.0
